Added subdirectories with placeholders
I didn't add anything but config/ doc/ and include/, because I wasn't
sure if those would be populated with human generated files or not.
I figured that directories populated with only computer generated files
shouldn't be tracked by git. If we need to add more later, we can.

Also, when we populate these directories, we should remove the
.placeholder files.