[issue 7] Bash usage on Windows
I consider issue 7 closed with this commit, although the requirement
that Bash scripts be compatible with Bash 2.1 hasn't been documented
yet.  I'm assuming that we'll get back to that when someone actually
decides to write a Bash script, since Python is much more convenient
as a cross-platform scripting language anyway.