Remove system installation for Windows from setup
Made %HOME% the default install location for rsshell in Windows. This
is not a system install as intended but at least it works. A complete
solution in the future might require a full-blown MSI installer.

A working system installer for Windows was almost completed, but I
wasn't able to solve a problem with file/directory permissions. I was
running setup.py from an administrator account, and the script was
apparently sufficiently privileged to write to %PROGRAMFILES% since
the check on line 144 didn't fire. However, on line 189 the program
was interrupted with an IOError. Somehow, having write permissions to
a directory in Windows doesn't mean that one can create
subdirectories.

I tried manually creating the subdirectory (%PROGRAMFILES%\Red Spider
Project) and then running the script again, but it would just get
stuck on the next creation at line 191. I've attached the traceback at
the end of this commit message.

I've created the branch 'jgonggrijp/setup-windows' so that the old
state of setup.py can be permanently accessed. Note that the line
numbers in my explanation above, as well as those in the traceback
below, refer to the state in jgonggrijp/setup-windows, not to the
state in the present commit.

The install_rsshell_windows function has been moved past the
invocation of main.

Traceback (most recent call last):
  File "C:\Users\Julian\Documents\RedSpider\setup.py", line 314, in <module>
    main()
  File "C:\Users\Julian\Documents\RedSpider\setup.py", line 53, in main
    install()
  File "C:\Users\Julian\Documents\RedSpider\setup.py", line 99, in install
    install_rsshell()
  File "C:\Users\Julian\Documents\RedSpider\setup.py", line 132, in install_rsshell
    install_rsshell_windows(src_file, fname)
  File "C:\Users\Julian\Documents\RedSpider\setup.py", line 166, in install_rsshell_windows
    rsshell_install_finish(src_file, rsshell_target_dir, fname)
  File "C:\Users\Julian\Documents\RedSpider\setup.py", line 191, in rsshell_install_finish
    copy2(src_file, bin_file)
  File "C:\Python27\lib\shutil.py", line 128, in copy2
    copyfile(src, dst)
  File "C:\Python27\lib\shutil.py", line 83, in copyfile
    with open(dst, 'wb') as fdst:
IOError: [Errno 13] Permission denied: 'C:\\Program Files\\Red Spider Project\\rsshell.py'