[issue 11] Option 4: install rsshell internally but do add to PATH
While considering the option to install rsshell internally in the
project root on Windows (instead of system-wide), I realized that this
would be the best option anyway for all platforms. It improves
consistency across platforms, it simplifies the program logic of
setup.py, it removes the need to run setup.py with administrative
rights and it allows multiple users on the same computer to install
different version of rsshell.

The only real difference with previous versions on the master branch
is that a special subdirectory is reserved for externally visible
executables: extbin, with a matching extension of the user's PATH
settings. This directory has also been added to the .gitignore file.