Save RED_SPIDER_ROOT to user environment as well
We're appending /extbin to the user's PATH anyway, so it makes sense
to do the same thing with RED_SPIDER_ROOT. This removes the need to
create ~/.config/xkcdRedSpider (POSIX) or %APPDATA%\xkcdRedSpider
(Windows).

From now on, we'll need to modify only one location that is not within
the project root: ~/.profile on POSIX or HKEY_CURRENT_USER in the
Registry on Windows.