Enable calling 'rsshell' without the '.py' extension on Windows
If setup.py is run on Windows and it detects that '.py' is not
included in the PATHEXT environment variable, it will insert a small
additional batch script in ROOT\extbin with the name 'rsshell.cmd'.
This script simply calls extbin\rsshell.py. Since '.cmd' is always in
PATHEXT, rsshell can now be called without needing to type the
extension.