Replaced py.test with unittest
Since it is apparently rather difficult to make py.test work with
python3 in a windows environment, I have replaced it with the
python-unittest framework, which is in the standard library and should
theoretically work from anywhere.

I have replaced assert statments with unnittest methods. This is the
idiomatic way to make assertions in unittest, which allows for magical
introspection and more informative failure reports.

I have updated the Readme with new testing instructions.