Use shell and enquote argv on rsshell argument call
This is almost back to the old approach for the single-command call:
manually concatenate the arguments with spaces and let the system
shell intervene. The only differences that survive are the use of
subprocess.call instead of os.system and the enquotation of the
arguments in order to preserve multiword arguments.

The use of the shell is consistent, because rsshell uses the shell for
interactive mode as well. It's also necessary to allow calling Red
Spider commands with the usual environment (i.e. being able to omit
path and extension), and it might be the only way to make the argument
version of rsshell work on Windows at all (see first reaction to #21).

The manual concatenation is a bit inelegant, but it seems to be the
only way (at least on Mac OS X) to make subprocess.call correctly pass
subsequent arguments when shell=True.